Namespace test message model
