+++ app/models/git/exercise_reader.rb
@@ -31,7 +31,7 @@ class Git::ExerciseReader
 
   def test_messages
     Array(exercise_config[:test_messages]).
-      map { |message| TestMessage.from_file(message) }
+      map { |message| Git::TestMessage.from_file(message) }
   end
 
   def tests
